+++ src/server.rs
@@ -266,7 +266,7 @@ impl tower_lsp::LanguageServer for PathServer {
         .await;
 
         // completion
-        let completion_config = self.get_config().await.completion;
+        let config = self.get_config().await;
         let Ok(file_path) = url_to_path(&params.text_document_position.text_document.uri) else {
             warn(format!(
                 "Failed to convert URI to file path: {}",
@@ -277,8 +277,7 @@ impl tower_lsp::LanguageServer for PathServer {
         };
         let workspace_roots = self.workspace_roots.read().await;
         let completions =
-            providers::complete(&raw_path, &workspace_roots, &file_path, &completion_config)
-                .await?;
+            providers::complete(&raw_path, &workspace_roots, &file_path, &config).await?;
         debug(format!(
             "<Completion> Generated completions: {}",
             completions.len()
@@ -310,8 +309,11 @@ impl tower_lsp::LanguageServer for PathServer {
             warn(format!("Document not found: {}", path.display())).await;
             return Ok(None);
         };
+        let config = self.get_config().await;
+        let workspace_roots = self.workspace_roots.read().await;
 
-        let links = providers::provide_document_links(doc, &path).await?;
+        let links =
+            providers::provide_document_links(doc, &path, &config, &workspace_roots).await?;
         debug(format!(
             "<Document Link> Generated document links: {}",
             links.len()
@@ -349,8 +351,12 @@ impl tower_lsp::LanguageServer for PathServer {
             .await;
             return Ok(None);
         };
+        let config = self.get_config().await;
+        let workspace_roots = self.workspace_roots.read().await;
 
-        let definition = providers::provide_definition(doc, line, character, &path).await?;
+        let definition =
+            providers::provide_definition(doc, &path, line, character, &config, &workspace_roots)
+                .await?;
         if let Some(definition) = &definition {
             let lsp_types::GotoDefinitionResponse::Scalar(definition) = &definition else {
                 unreachable!("Definition is not a scalar");