+++ CHANGELOG.md
@@ -9,11 +9,15 @@ All notable changes to the Path Server will be documented in this file.
 - **Core**: Added **Document Links** provider support in supported editors. (zed's api does not supported it for now)
     - Automatically detects and underlines valid file paths in the editor.
     - Making paths clickable and allowing users to jump directly to the target file.
-    - *Note: Currently not support in Zed due to API limitations*
+    - *Note: Currently not supported in **Zed** as it does not yet implement the LSP Document Link feature.*
 - **Core**: Added **Go to Definition** provider support.
     - Enables standard "Go to Definition" functionality for string-based file paths.
     - Users can now use editor shortcuts (e.g., `Cmd/Ctrl + Click` or `F12`) to instantly open the file referenced by a path.
+- **Core**: Added config entry: `path-server.highlight.enable` to control highlighting of file paths in editor.
 - **VS Code**: Add command `Path Server: Restart Server` to restart the Path Server.
+
+### Changed
+- **Core**: Promote config entry from `path-server.completion.basePath` to `path-server.basePath`. Now this configuration is used for both completion and other features.
 
 ## [0.3.0] 2026-03-06
 ### Fixed